backend/requirements.txt
requests
python-dotenv
pandas
flask
flask-cors
croniter
Pillow
imagehash
imutils
opencv-python-headless
waitress
m3u8
SQLAlchemy
